% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_prep.R
\name{azure_prep}
\alias{azure_prep}
\title{Prepare Azure environment and create SQLite databases}
\usage{
azure_prep()
}
\value{
No return value, called for side effects
(sets environment variables and creates SQLite databases)
}
\description{
This function sets up the necessary environment variables for running the
application in Azure and creates the required SQLite database files and
tables.
}
\keyword{internal}
