% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cases_year}
\alias{cases_year}
\title{Measles cases across the world}
\format{
## `cases_year`
A data frame with 2,382 rows and 19 variables:
\describe{
\item{year}{Year of observation}
\item{country}{Country name}
\item{iso3c}{ISO 3166-1 alpha-3 country code}
\item{region}{Region name}
...
}
}
\source{
<https://github.com/rfordatascience/tidytuesday/tree/main/data/2025/2025-06-24>
}
\usage{
cases_year
}
\description{
Taken from https://github.com/rfordatascience/tidytuesday/tree/main/data/2025/2025-06-24
A dataset containing the number of reported measles cases per country per year.
}
\keyword{datasets}
