% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_prep.R
\name{create_sqlite_db}
\alias{create_sqlite_db}
\title{Create log and autoreport sqlite database files and tables}
\usage{
create_sqlite_db()
}
\value{
No return value, called for side effects
(creates SQLite database files and tables)
}
\description{
This function will create two files that can be used as database
when we do not have access to database infrastructure.
It will create the necessary tables within these files.
Please be aware that this will delete any existing tables
with the same names!
}
