% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_prep.R
\name{define_azure_env}
\alias{define_azure_env}
\title{Define Azure environment variables}
\usage{
define_azure_env()
}
\value{
No return value, called for side effects (sets environment variables)
}
\description{
This function sets the necessary environment variables for running the
application in Azure.
}
\keyword{internal}
