% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_fordeling.R
\name{lag_fordeling_plot}
\alias{lag_fordeling_plot}
\title{Plot: fordeling}
\usage{
lag_fordeling_plot(
  data,
  var,
  valg_sammenligne_grupper,
  var_sammenligne,
  y_lab_size = 12,
  x_lab_size = 12
)
}
\arguments{
\item{data}{datafil som har vært gjennom forbered_data_fordeling() og utvalg_fordeling()}

\item{var}{variabel valgt av bruker i ui-delen}

\item{valg_sammenligne_grupper}{"ja" eller "nei" valg av bruker i ui-delen}

\item{var_sammenligne}{variabel for sammenligning valg av bruker i ui-delen}

\item{y_lab_size}{størrelse på y-akse tekst}

\item{x_lab_size}{størrelse på x-akse tekst}
}
\value{
ggplot2-object som viser fordeling pr. gruppe
}
\description{
Plot: fordeling
}
