% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{licorice_gargle}
\alias{licorice_gargle}
\title{Licorice gargling data}
\format{
## `licorice_gargle`
A data frame with 235 rows and 19 variables:
\describe{
...
}
}
\usage{
licorice_gargle
}
\description{
Licorice gargling data
}
\keyword{datasets}
