% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_andel_plot.R
\name{mod_andeler_ui}
\alias{mod_andeler_ui}
\title{Shiny module providing GUI and server logic for the Andeler tab}
\usage{
mod_andeler_ui(id)
}
\arguments{
\item{id}{Character string module namespace}
}
\value{
An shiny app ui object
}
\description{
Shiny module providing GUI and server logic for the Andeler tab
}
