% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fordeling_plot.R
\name{mod_fordeling_plot_server}
\alias{mod_fordeling_plot_server}
\title{Server fordeling}
\usage{
mod_fordeling_plot_server(id, data)
}
\arguments{
\item{id}{Character string module namespace}

\item{data}{Data frame containing the data to be plotted.}
}
\description{
Server fordeling
}
