% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_over_tid.R
\name{mod_over_tid_server}
\alias{mod_over_tid_server}
\title{Server fordeling}
\usage{
mod_over_tid_server(id, data, indicator_meta)
}
\arguments{
\item{id}{Character string module namespace}

\item{data}{Data frame containing the data to be plotted.}

\item{indicator_meta}{Data frame containing metadata for the indicators.}
}
\description{
Server fordeling
}
