% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_over_tid.R
\name{mod_over_tid_UI}
\alias{mod_over_tid_UI}
\title{Shiny module providing GUI and server logic for the plot tab}
\usage{
mod_over_tid_UI(id)
}
\arguments{
\item{id}{Character string module namespace}
}
\value{
An shiny app ui object
}
\description{
Shiny module providing GUI and server logic for the plot tab
}
