% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_fordeling.R
\name{plotGruppertFordeling}
\alias{plotGruppertFordeling}
\title{Plot: gruppert fordeling}
\usage{
plotGruppertFordeling(
  data,
  fordelingsVariabel = NULL,
  binVariabel,
  binNavn = NULL,
  colors = NULL,
  totalLabel = "Totalt",
  yLabel = "Andel",
  fillLabel = NULL
)
}
\arguments{
\item{data}{datafil som har vært gjennom forbered_data_fordeling() og utvalg_fordeling()}

\item{fordelingsVariabel}{variabel for fordeling, NULL for totalt}

\item{binVariabel}{variabel for binning}

\item{binNavn}{optional mapping for bin values}

\item{colors}{optional color palette}

\item{totalLabel}{label for total row}

\item{yLabel}{label for y-axis}

\item{fillLabel}{label for fill legend}
}
\value{
ggplot2-object som viser fordeling pr. gruppe
}
\description{
Plot: gruppert fordeling
}
