% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_over_tid.R
\name{plotSPC}
\alias{plotSPC}
\title{Lag et SPC-diagram for valgt indikator og sykehus}
\usage{
plotSPC(data, title = NULL, subtitle = NULL)
}
\arguments{
\item{data}{Dataramme med kolonnene \code{year}, \code{unitName}, \code{var},
\code{denominator} og \code{ind_id}. Metadata som \code{title} og
\code{short_description} brukes dersom de finnes.}

\item{title}{Valgfri tittel for diagrammet. Hvis ikke spesifisert, brukes
standardtittel "SPC-diagram".}

\item{subtitle}{Valgfri undertittel for diagrammet. Hvis ikke spesifisert
brukes undertittel fra metadata dersom tilgjengelig, ellers ingen undertittel.}
}
\value{
Et ggplot-objekt med SPC-diagram for valgt indikator.
}
\description{
Lager et p-diagram med
\code{qicharts2} basert på indikatorverdier filtrert på valgt sykehus
og indikator.
}
