% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_utsendinger.R, R/mod_abonnement.R
\name{samlerapport1Fun}
\alias{samlerapport1Fun}
\title{Example report function 1}
\usage{
samlerapport1Fun(file = NULL, p1 = "Alder", p2 = 1, reshID = 999999)

samlerapport1Fun(file = NULL, p1 = "Alder", p2 = 1, reshID = 999999)
}
\value{
Path to the generated report file

Path to the generated report file
}
\description{
Example report function 1

Example report function 1
}
