% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_utsendinger.R
\name{samlerapport2Fun}
\alias{samlerapport2Fun}
\title{Example report function 2}
\usage{
samlerapport2Fun(file = NULL, p1 = "Alder", p2 = 1, reshID = 999999)
}
\arguments{
\item{file}{Path to the output file. If NULL, a temporary file will be created.}

\item{p1}{Parameter 1 (default: "Alder")}

\item{p2}{Parameter 2 (default: 1)}

\item{reshID}{ReshID (default: 999999)}
}
\value{
Path to the generated report file
}
\description{
Example report function 2
}
