% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_utsending.R
\name{utsending_ui}
\alias{utsending_ui}
\title{Shiny module providing GUI and server logic for the dispatch tab}
\usage{
utsending_ui(id)
}
\arguments{
\item{id}{Character string module namespace}
}
\value{
An shiny app ui object
}
\description{
Shiny module providing GUI and server logic for the dispatch tab
}
