% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_fordeling.R
\name{utvalg}
\alias{utvalg}
\title{Funksjon som gjør utvalg basert på  ui-valg}
\usage{
utvalg(data, alder1, alder2, roek)
}
\arguments{
\item{data}{datafil som utgangspunkt}

\item{alder1}{minste alder}

\item{alder2}{høyeste alder}

\item{roek_1_2_3}{røyking 1 = nå, 2 = før, 3 = aldri}
}
\value{
datafil der utvalg er gjort
}
\description{
Funksjon som gjør utvalg basert på  ui-valg
}
