% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_veiledning.R
\name{veiledning_ui}
\alias{veiledning_ui}
\title{Shiny module providing GUI and server logic for the intro tab}
\usage{
veiledning_ui(id)
}
\arguments{
\item{id}{Character string module namespace}
}
\value{
An shiny app ui object
}
\description{
Shiny module providing GUI and server logic for the intro tab
}
