% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_andel_plot.R
\name{mod_andeler_server}
\alias{mod_andeler_server}
\title{Server logic for andel plot}
\usage{
mod_andeler_server(id, data, indicator_meta)
}
\arguments{
\item{id}{Character string module namespace}

\item{data}{Data frame containing the data to be plotted.}
}
\value{
A Shiny app server object
}
\description{
Server logic for andel plot
}
