% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeAndelPlot.R
\name{PlotAndelerGrVar}
\alias{PlotAndelerGrVar}
\title{PlotAndelerGrVar}
\usage{
PlotAndelerGrVar(
  RegData,
  Variabel,
  grVar = "ShNavn",
  hovedgrTxt = "Total andel",
  kvalIndGrenser = NA,
  tittel = "tittel",
  utvalgTxt = "",
  Ngrense = 10,
  bestKvalInd = "lav",
  fargepalett = "BlaaOff",
  grtxt = "",
  titleSize = 20,
  subtitleSize = 15,
  legendSize = 12,
  axisTextSize = 12,
  nTicks = 5
)
}
\arguments{
\item{RegData}{A data frame containing the registry data to be used for plotting.}
}
\value{
A plot object visualizing the group-specific proportions or distributions.
}
\description{
This function generates a plot for group-specific proportions or distributions
             based on the provided registry data.
}
\details{
This function is designed to work with registry data and create visualizations
         that highlight group-specific metrics. The exact details of the plot depend on
         the structure of the input data and the specific implementation of the function.
}
