% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportProcessor.R
\name{reportProcessor}
\alias{reportProcessor}
\title{Common report processor for rapRegTemplate}
\usage{
reportProcessor(report, outputType = "pdf", var = "mpg", bins = 10)
}
\arguments{
\item{report}{The type of report to generate. Currently only "local_monthly" is supported.}

\item{outputType}{The format of the output file. Can be "html", "html_fragment", or "pdf".}

\item{var}{The variable to be used in the report.}

\item{bins}{The number of bins to use for the report.}
}
\value{
A character string with a path to where the produced file is located.
}
\description{
Makes reports for rapRegTemplate typically used for auto reports such as
subscriptions, dispatchments and bulletins.
}
