% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_abonnement.R
\name{samlerapport2Fun}
\alias{samlerapport2Fun}
\title{Example report function 2}
\usage{
samlerapport2Fun(file = NULL, p1 = "Alder", p2 = 1, reshID = 999999)
}
\value{
Path to the generated report file
}
\description{
Example report function 2
}
